+++ test/lib/deployment/transient-error.js
